pyproject.toml
[project]
name = "ild-mdt-bench"
version = "0.1.0"
description = "ILD-MDT 多智能体临床推理实验平台"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.30",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "langgraph>=0.2",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]
analysis = [
    "pandas>=2.0",
    "jupyter>=1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
